# openapi.yaml
openapi: 3.0.0
info:
  title: PB138 Study Planner API
  description: API specification for the PB138 Study Planning Application
  version: 1.0.0
  contact:
    name: PB138 Project
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pb138.example.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskPriority:
      type: string
      enum: [NONE, LOW, MEDIUM, HIGH, URGENT]
      description: Priority level of a task
    Task:
      type: object
      required:
        - id
        - userId
        - title
        - done
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the task
          example: 123
        userId:
          type: integer
          format: int64
          description: ID of the user who owns this task
          example: 1
        title:
          type: string
          description: Task title
          example: "Write essay on history"
          minLength: 1
        done:
          type: boolean
          description: Whether the task is completed
          default: false
          example: false
        subjectId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated subject, if any
          example: 42
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated study plan, if any
          example: 5
        favorite:
          type: boolean
          description: Whether the task is marked as favorite
          default: false
          example: false
        tag:
          type: string
          nullable: true
          description: Custom tag or label for the task
          example: "homework"
        deadline:
          type: string
          format: date-time
          nullable: true
          description: Deadline for completing the task (ISO 8601 format)
          example: "2026-05-15T23:59:59.000Z"
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Soft delete timestamp (null if not deleted)
          example: null
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Task title
          minLength: 1
          example: "Study for exam"
        done:
          type: boolean
          description: Whether the task is completed
          default: false
          example: false
        subjectId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated subject
          example: 42
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated study plan
          example: 5
        favorite:
          type: boolean
          description: Mark as favorite
          default: false
          example: false
        tag:
          type: string
          nullable: true
          description: Custom tag or priority name
          example: "urgent"
        priority:
          $ref: '#/components/schemas/TaskPriority'
          description: Task priority (alternative to tag)
          example: "HIGH"
        deadline:
          type: string
          format: date
          nullable: true
          description: Deadline date (ISO 8601 format)
          example: "2026-05-15"
...
    BulkTaskRequest:
      type: object
      required:
        - tasks
      properties:
        tasks:
          type: array
          description: Array of tasks to upsert
          minItems: 1
          items:
            type: object
            required:
              - id
              - title
              - done
            properties:
              id:
                type: integer
                format: int64
                example: 123
              title:
                type: string
                minLength: 1
                example: "Task title"
              done:
                type: boolean
                example: false
              subjectId:
                type: integer
                format: int64
                nullable: true
                example: 42
    PaginatedTasks:
      type: object
      required:
        - data
        - hasMore
        - nextCursor
        - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        hasMore:
          type: boolean
          description: Whether there are more results available
          example: false
        nextCursor:
          type: string
          nullable: true
          description: Cursor for the next page
          example: "456"
        limit:
          type: integer
          description: Number of items returned
          example: 30
    Subject:
      type: object
      required:
        - id
        - userId
        - name
        - teacher
        - code
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the subject
          example: 42
        userId:
          type: integer
          format: int64
          description: ID of the user who owns this subject
          example: 1
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated study plan
          example: 5
        name:
          type: string
          description: Subject name
          example: "Linear Algebra"
        teacher:
          type: string
          description: Teacher's name
          example: "Prof. Smith"
        code:
          type: string
          description: Subject code (uppercase)
          example: "LA101"
          pattern: "^[A-Z0-9]+$"
        isShared:
          type: boolean
          description: Whether the subject is shared with other users
          default: false
          example: false
        archived:
          type: boolean
          description: Whether the subject is archived (soft deleted)
          example: false
        files:
          type: integer
          description: Count of files associated with this subject
          example: 3
        tasks:
          type: integer
          description: Count of tasks associated with this subject
          example: 5
        events:
          type: integer
          description: Count of events associated with this subject
          example: 2
        notes:
          type: integer
          description: Count of lesson notes associated with this subject
          example: 4
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-05-06T14:20:00.000Z"
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Soft delete timestamp (null if not deleted)
          example: null
    CreateSubjectRequest:
      type: object
      required:
        - name
        - teacher
        - code
      properties:
        name:
          type: string
          minLength: 1
          description: Subject name
          example: "Linear Algebra"
        teacher:
          type: string
          minLength: 1
          description: Teacher's name
          example: "Prof. Smith"
        code:
          type: string
          minLength: 1
          description: Subject code (will be uppercase)
          example: "LA101"
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated study plan
          example: 5
        isShared:
          type: boolean
          description: Whether to share with other users
          default: false
          example: false
...
    PaginatedSubjects:
      type: object
      required:
        - data
        - hasMore
        - nextCursor
        - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Subject'
        hasMore:
          type: boolean
          example: false
        nextCursor:
          type: string
          nullable: true
          example: "100"
        limit:
          type: integer
          example: 20
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
    BulkTasksResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/Task'
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "Neplatne ID ukolu."
security:
  - bearerAuth: []
paths:
  /api/tasks:
    get:
      summary: List tasks with optional filtering and pagination
      tags:
        - Tasks
      operationId: listTasks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 30
          description: Number of items per page (max 200)
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination (ID of the last item from previous page)
        - name: paginated
          in: query
          schema:
            type: boolean
          description: Enable pagination mode (true/false)
        - name: subjectId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by subject ID
        - name: studyPlanId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by study plan ID
        - name: done
          in: query
          schema:
            type: boolean
          description: Filter by completion status (true/false)
        - name: favorite
          in: query
          schema:
            type: boolean
          description: Filter by favorite status (true/false)
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag/label
        - name: search
          in: query
          schema:
            type: string
          description: Search in task titles (case-insensitive)
        - name: deadlineFrom
          in: query
          schema:
            type: string
            format: date
          description: Filter tasks with deadline from this date (ISO 8601)
        - name: deadlineTo
          in: query
          schema:
            type: string
            format: date
          description: Filter tasks with deadline until this date (ISO 8601)
        - name: includeDeleted
          in: query
          schema:
            type: boolean
          description: Include soft-deleted tasks (default false)
      responses:
        '200':
          description: List of tasks (with or without pagination)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Task'
                    description: Regular list when pagination is not enabled
                  - $ref: '#/components/schemas/PaginatedTasks'
                    description: Paginated response when pagination is enabled
              examples:
                unpaginated:
                  value:
                    - id: 1
                      userId: 1
                      title: "Study for exam"
                      done: false
                      subjectId: 42
                      studyPlanId: 5
                      favorite: false
                      tag: "urgent"
                      deadline: "2026-05-15T23:59:59.000Z"
                      deletedAt: null
                paginated:
                  value:
                    data:
                      - id: 1
                        userId: 1
                        title: "Study for exam"
                        done: false
                        subjectId: 42
                        studyPlanId: 5
                        favorite: false
                        tag: "urgent"
                        deadline: "2026-05-15T23:59:59.000Z"
                        deletedAt: null
                    hasMore: true
                    nextCursor: "42"
                    limit: 30
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new task
      tags:
        - Tasks
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              simple:
                summary: Simple task
                value:
                  title: "Read chapter 5"
              detailed:
                summary: Task with all fields
                value:
                  title: "Solve problem set"
                  done: false
                  subjectId: 42
                  studyPlanId: 5
                  favorite: true
                  tag: "homework"
                  deadline: "2026-05-20"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Bulk update/upsert tasks
      description: |
        Upsert (update or insert) multiple tasks at once.
        - Creates tasks with IDs that don't exist
        - Updates existing tasks
        - Soft-deletes tasks not in the request
      tags:
        - Tasks
      operationId: bulkUpdateTasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkTaskRequest'
            example:
              tasks:
                - id: 1
                  title: "Updated task"
                  done: true
                  subjectId: 42
                - id: 2
                  title: "Another task"
                  done: false
                  subjectId: null
      responses:
        '200':
          description: Tasks upserted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkTasksResponse'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/subjects:
    get:
      summary: List subjects with optional filtering and pagination
      description: |
        Retrieve subjects. Public users can only see shared subjects.
        Registered users can see their own subjects, shared subjects, and subjects in shared study plans.
      tags:
        - Subjects
      operationId: listSubjects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination (ID of the last item from previous page)
        - name: paginated
          in: query
          schema:
            type: boolean
          description: Enable pagination mode (true/false)
        - name: studyPlanId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by study plan ID
        - name: includeDeleted
          in: query
          schema:
            type: boolean
          description: Include archived subjects (default false)
      responses:
        '200':
          description: List of subjects (with or without pagination)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Subject'
                    description: Regular list when pagination is not enabled
                  - $ref: '#/components/schemas/PaginatedSubjects'
                    description: Paginated response when pagination is enabled
              examples:
                unpaginated:
                  value:
                    - id: 42
                      userId: 1
                      studyPlanId: 5
                      name: "Linear Algebra"
                      teacher: "Prof. Smith"
                      code: "LA101"
                      isShared: false
                      archived: false
                      deletedAt: null
                      files: 3
                      tasks: 5
                      events: 2
                      notes: 4
                      createdAt: "2026-01-15T10:30:00.000Z"
                      updatedAt: "2026-05-06T14:20:00.000Z"
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new subject
      description: |
        Create a subject. If studyPlanId is specified, the authenticated user must be
        the plan owner or a contributor to create subjects in it.
      tags:
        - Subjects
      operationId: createSubject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubjectRequest'
            examples:
              simple:
                summary: Simple subject
                value:
                  name: "Linear Algebra"
                  teacher: "Prof. Smith"
                  code: "LA101"
              withPlan:
                summary: Subject in a study plan
                value:
                  name: "Linear Algebra"
                  teacher: "Prof. Smith"
                  code: "LA101"
                  studyPlanId: 5
                  isShared: true
      responses:
        '201':
          description: Subject created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - no permission to create in this study plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Study plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...